#18289: Move yolov8s_world from experimental to demos folder
