+++ models/demos/yolov8s_world/reference/yolov8s_world.py
@@ -1,10 +1,10 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 
 # SPDX-License-Identifier: Apache-2.0
 
 import torch
-from torch import nn
 import torch.nn.functional as F
+from torch import nn
 
 
 def autopad(k, p=None, d=1):  # kernel, padding, dilation
