+++ models/demos/yolov8s_world/tt/ttnn_yolov8s_world_utils.py
@@ -1,32 +1,31 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 
 # SPDX-License-Identifier: Apache-2.0
 
+import math
 import os
 from pathlib import Path
 
 import torch
 import torch.nn as nn
-import math
 from loguru import logger
-
-import ttnn
 from ttnn.model_preprocessing import (
     ParameterDict,
     ParameterList,
-    preprocess_linear_weight,
-    preprocess_linear_bias,
     preprocess_layernorm_parameter,
+    preprocess_linear_bias,
+    preprocess_linear_weight,
 )
 
-from models.experimental.yolov8s_world.reference.yolov8s_world import (
-    Conv,
-    C2f,
+import ttnn
+from models.demos.yolov8s_world.reference.yolov8s_world import (
     SPPF,
+    C2f,
     C2fAttn,
-    WorldModel,
-    WorldDetect,
+    Conv,
     ImagePoolingAttn,
+    WorldDetect,
+    WorldModel,
 )
