+++ tests/ttnn/integration_tests/yolov8s_world/test_ttnn_yolov8s_world.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 
 # SPDX-License-Identifier: Apache-2.0
 
@@ -9,8 +9,8 @@ from loguru import logger
 import ttnn
 from tests.ttnn.utils_for_testing import assert_with_pcc
 from models.utility_functions import run_for_wormhole_b0
-from models.experimental.yolov8s_world.reference import yolov8s_world
-from models.experimental.yolov8s_world.tt.ttnn_yolov8s_world import (
+from models.demos.yolov8s_world.reference import yolov8s_world
+from models.demos.yolov8s_world.tt.ttnn_yolov8s_world import (
     TtConv,
     TtC2f,
     TtSPPF,
@@ -22,7 +22,7 @@ from models.experimental.yolov8s_world.tt.ttnn_yolov8s_world import (
     TtYOLOWorld,
     TtImagePoolingAttn,
 )
-from models.experimental.yolov8s_world.tt.ttnn_yolov8s_world_utils import (
+from models.demos.yolov8s_world.tt.ttnn_yolov8s_world_utils import (
     create_custom_preprocessor,
     attempt_load,
     move_to_device,
